Add the ability to specify a Charset when reading property files
Starting with Java 1.6, Properties has a .load() method accepting a Reader as an
argument. By default, the method using an InputStream as an argument will try
and decode the file using ISO-8859-1.

Change the code of .loadProperties() to use a Reader instead, and add the
ability to change the Charset used when reading files.